test: Refactored TF unittests for datasets
