+++ test/conftest.py
@@ -1,5 +1,6 @@
 import pytest
 import os
+import json
 import requests
 from io import BytesIO
 
@@ -43,3 +44,77 @@ def mock_image_folder(mock_image_stream, tmpdir_factory):
         with open(fn, 'wb') as f:
             f.write(file.getbuffer())
     return str(folder)
+
+
+@pytest.fixture(scope="session")
+def mock_detection_label(tmpdir_factory):
+    folder = tmpdir_factory.mktemp("labels")
+    label = {
+        "boxes_1": [[[1, 2], [1, 3], [2, 1], [2, 3]], [[10, 20], [10, 30], [20, 10], [20, 30]]],
+        "boxes_2": [[[3, 2], [3, 3], [4, 1], [4, 3]], [[30, 20], [30, 30], [40, 10], [40, 30]]],
+        "boxes_3": [[[1, 5], [1, 8], [2, 5], [2, 8]]],
+    }
+    for i in range(5):
+        fn = folder.join("mock_image_file_" + str(i) + ".jpeg.json")
+        with open(fn, 'w') as f:
+            json.dump(label, f)
+    return str(folder)
+
+
+@pytest.fixture(scope="session")
+def mock_recognition_label(tmpdir_factory):
+    label_file = tmpdir_factory.mktemp("labels").join("labels.json")
+    label = {
+        "mock_image_file_0.jpeg": "I",
+        "mock_image_file_1.jpeg": "am",
+        "mock_image_file_2.jpeg": "a",
+        "mock_image_file_3.jpeg": "jedi",
+        "mock_image_file_4.jpeg": "!",
+    }
+    with open(label_file, 'w') as f:
+        json.dump(label, f)
+    return str(label_file)
+
+
+@pytest.fixture(scope="session")
+def mock_ocrdataset(tmpdir_factory, mock_image_stream):
+    root = tmpdir_factory.mktemp("dataset")
+    label_file = root.join("labels.json")
+    label = [
+        {
+            "raw-archive-filepath": "mock_image_file_0.jpg",
+            "coordinates": [[[0.1, 0.2], [0.2, 0.2], [0.2, 0.3], [0.1, 0.3]]],
+            "string": ['I']
+        },
+        {
+            "raw-archive-filepath": "mock_image_file_1.jpg",
+            "coordinates": [[[0.1, 0.2], [0.2, 0.2], [0.2, 0.3], [0.1, 0.3]]],
+            "string": ['am']
+        },
+        {
+            "raw-archive-filepath": "mock_image_file_2.jpg",
+            "coordinates": [[[0.1, 0.2], [0.2, 0.2], [0.2, 0.3], [0.1, 0.3]]],
+            "string": ['a']
+        },
+        {
+            "raw-archive-filepath": "mock_image_file_3.jpg",
+            "coordinates": [[[0.1, 0.2], [0.2, 0.2], [0.2, 0.3], [0.1, 0.3]]],
+            "string": ['jedi']
+        },
+        {
+            "raw-archive-filepath": "mock_image_file_4.jpg",
+            "coordinates": [[[0.1, 0.2], [0.2, 0.2], [0.2, 0.3], [0.1, 0.3]]],
+            "string": ['!']
+        }
+    ]
+    with open(label_file, 'w') as f:
+        json.dump(label, f)
+
+    file = BytesIO(mock_image_stream)
+    image_folder = tmpdir_factory.mktemp("images")
+    for i in range(5):
+        fn = image_folder.join(f"mock_image_file_{i}.jpg")
+        with open(fn, 'wb') as f:
+            f.write(file.getbuffer())
+
+    return str(image_folder), str(label_file)
